The CYRKeyboardButton shows the popup above the key except in the top row where there's not enough room. Then the popup is overlaid.
However, the logic for which keys are in the top row was hardcoded only to an English QWERTY layout. For a layout like AZERTY it would assume the "A" was not a top row key and the popup would not fit.
Now get the info about top row from when the keys are being laid out.

NB perhaps even this is a hack: really the question is: will the pop up fit within the bounding frame of the keyboard?

Also fix error where special keys didn't switch to dark mode.