Eliminate the options key from the bottom row. It's now one of the options if you long press on the globe key.
Along for the ride: extensive refactoring of the layout logic for the bottom row of the keyboard since it had all sorts of baked-in assumptions, e.g. it assumed there was always going to be an option key (which it called a mic key) to the left of the space bar.

Closes #21.